// src/app/shared/enums/http-entities.enum.ts
export enum HttpEntitiesEnum {
    BRANCHES = 'sucursal',
    BRAND = 'marca',
    CATEGORY_SUB = 'categoria_sub',
    PRODUCTS = 'productos',
    PROMOTIONS = 'promociones',
    CUSTOMERS = 'clientes',
    LOCATIONS = 'ubicacion',
    ORDERS = 'ordenes',
    CART = 'carrito',
    PAYMENTS = 'pagos',
    CATEGORY = 'categoria',
    AUTH = 'auth',
    COUPONS = 'cupones',
    PETS = 'mascotas',
    MEDIA = 'media',
    INVOICE = 'facturacion',
    DELIVERY_COST = 'tarifa_envio',
}